svm: move callback into new module
