+++ svm/src/lib.rs
@@ -7,6 +7,7 @@ pub mod account_rent_state;
 pub mod message_processor;
 pub mod transaction_account_state_info;
 pub mod transaction_error_metrics;
+pub mod transaction_processing_callback;
 pub mod transaction_processor;
 pub mod transaction_results;
